extract `portfolio.pnl`
