+++ alphaswarm/services/portfolio/__init__.py
@@ -1,4 +1,5 @@
 from .portfolio import Portfolio
-from .portfolio_base import PortfolioBase, PortfolioSwap, PortfolioPNL, PortfolioPNLDetail, PnlMode
+from .portfolio_base import PortfolioBase, PortfolioSwap
 from .portfolio_evm import PortfolioEvm
+from .portfolio_pnl import PortfolioPNL, PortfolioPNLDetail, PnlMode
 from .portfolio_solana import PortfolioSolana
